actualizaciones negocion 27-07-23

+++ frontend/components/ClienteIndividual.jsx
@@ -9,24 +9,41 @@ const ClienteIndividual = ({ cliente }) => {
 
   //Función para eliminar el cliente
   const navegar = useNavigate();
-  function eliminarCliente() {
-    fetch(`http://localhost:4000/api/cliente/eliminarcliente/${_id}`, {  // Corrección en la ruta
-      method: 'DELETE',
-      headers: {
-        'Content-Type': 'application/json'
+  const eliminarCliente = () => {
+    swal({
+      title: "Eliminar",
+      text: "¿Estás seguro de eliminar el registro?",
+      icon: "warning",
+      buttons: {
+        cancel: "NO",
+        confirm: "SI"
+      },
+      dangerMode: true
+    }).then(isConfirmed => {
+      if (isConfirmed) {
+        fetch(`http://localhost:4000/api/cliente/eliminarcliente/${_id}`, {
+          method: 'DELETE',
+          headers: {
+            'Content-Type': 'application/json'
+          }
+        })
+          .then(res => res.json())
+          .then(data => {
+            console.log(data);
+            swal({
+              title: "Cliente eliminado con éxito",
+              icon: "success"
+            }).then(() => {
+              navegar(0);
+            });
+          })
+          .catch(error => {
+            console.error('Error:', error);
+            alert('Error al eliminar el cliente');
+          });
       }
-    })
-      .then(res => res.json())
-      .then(data => {
-        // console.log(data);
-        alert(data.message); // Mostrar el mensaje específico del objeto
-        navegar(0);
-      })
-      .catch(error => {
-        console.error('Error:', error);
-        alert('Error al eliminar el cliente');
-      });
-  }
+    });
+  };
 
   const [mostrarDetalles, setMostrarDetalles] = useState(false); // Estado para controlar la ventana emergente
 
@@ -38,7 +55,7 @@ const ClienteIndividual = ({ cliente }) => {
   const customStyles = {
     content: {
       width: '700px',
-      height: '570px',
+      height: '590px',
       margin: 'auto',
       borderRadius: '10px',
       padding: '20px',
@@ -57,8 +74,6 @@ const ClienteIndividual = ({ cliente }) => {
   }
 
   return (
-
-    
     <tr>
       <td>{cliente.cedula}</td>
       <td>{cliente.nombre}</td>
@@ -85,43 +100,43 @@ const ClienteIndividual = ({ cliente }) => {
         <br />
         <h2 style={{ textAlign: 'center', color: '#032770' }}>DETALLE CLIENTE</h2>
         <br />
-        <table className="table table-hover mb-5 table-bordered" style={{ maxWidth: 800, border: "2px solid blue"}}>
+        <table className="table table-hover mb-5 table-bordered" style={{ maxWidth: 800, border: "2px solid blue" }}>
           <tbody>
             <tr>
               <th scope="row" style={{ backgroundColor: "#032770", color: 'white' }}>Cédula</th>
-              <td style={{color: '#032770'}}>{cliente.cedula}</td>
+              <td style={{ color: '#032770' }}>{cliente.cedula}</td>
             </tr>
             <tr>
               <th scope="row" style={{ backgroundColor: "#032770", color: 'white' }}>Nombre</th>
-              <td style={{color: '#032770'}}>{cliente.nombre}</td>
+              <td style={{ color: '#032770' }}>{cliente.nombre}</td>
             </tr>
             <tr>
               <th scope="row" style={{ backgroundColor: "#032770", color: 'white' }}>Dirección</th>
-              <td style={{color: '#032770'}}>{cliente.direccion}</td>
+              <td style={{ color: '#032770' }}>{cliente.direccion}</td>
             </tr>
             <tr>
               <th scope="row" style={{ backgroundColor: "#032770", color: 'white' }}>Teléfono</th>
-              <td style={{color: '#032770'}}>{cliente.telefono}</td>
+              <td style={{ color: '#032770' }}>{cliente.telefono}</td>
             </tr>
             <tr>
               <th scope="row" style={{ backgroundColor: "#032770", color: 'white' }}>Email</th>
-              <td style={{color: '#032770'}}>{cliente.email}</td>
+              <td style={{ color: '#032770' }}>{cliente.email}</td>
             </tr>
             <tr>
               <th scope="row" style={{ backgroundColor: "#032770", color: 'white' }}>Cédula Codeudor</th>
-              <td style={{color: '#032770'}}>{cliente.cedulaCodeudor}</td>
+              <td style={{ color: '#032770' }}>{cliente.cedulaCodeudor}</td>
             </tr>
             <tr>
               <th scope="row" style={{ backgroundColor: "#032770", color: 'white' }}>Nombre Codeudor</th>
-              <td style={{color: '#032770'}}>{cliente.nombreCodeudor}</td>
+              <td style={{ color: '#032770' }}>{cliente.nombreCodeudor}</td>
             </tr>
             <tr>
               <th scope="row" style={{ backgroundColor: "#032770", color: 'white' }}>Teléfono Codeudor</th>
-              <td style={{color: '#032770'}}>{cliente.telefonoCodeudor}</td>
+              <td style={{ color: '#032770' }}>{cliente.telefonoCodeudor}</td>
             </tr>
             <tr>
               <th scope="row" style={{ backgroundColor: "#032770", color: 'white' }}>Grupo</th>
-              <td style={{color: '#032770'}}>{cliente.grupo}</td>
+              <td style={{ color: '#032770' }}>{cliente.grupo}</td>
             </tr>
           </tbody>
         </table>
