+++ frontend/components/NegociacionIndividual.jsx
@@ -36,50 +36,65 @@ const NegociacionIndividual = ({ negociacion }) => {
 
   const customStyles = {
     content: {
-      width: '600px',
-      height: '600px',
+      width: '700px',
+      height: '590px',
       margin: 'auto',
-      borderRadius: '50px',
+      borderRadius: '10px',
       padding: '20px',
       display: 'flex',
       flexDirection: 'column',
-      // justifyContent: 'center',
     },
   };
 
   const customStyles2 = {
     content: {
       width: '700px',
-      height: '500px',
+      height: '590px',
       margin: 'auto',
-      borderRadius: '50px',
+      borderRadius: '10px',
       padding: '20px',
       display: 'flex',
       flexDirection: 'column',
-      // justifyContent: 'center',
     },
   };
 
   //Función para eliminar la negociacion
   const navegar = useNavigate()
   function eliminarnegociacion() {
-    fetch(`http://localhost:4000/api/negociacion/eliminarnegociacion/${_id}`, {
-      method: 'DELETE',
-      headers: {
-        'Content-Type': 'application/json'
+    swal({
+      title: "Eliminar",
+      text: "¿Estás seguro de eliminar el registro?",
+      icon: "warning",
+      buttons: {
+        cancel: "NO",
+        confirm: "SI"
+      },
+      dangerMode: true
+    }).then(isConfirmed => {
+      if (isConfirmed) {
+        fetch(`http://localhost:4000/api/negociacion/eliminarnegociacion/${_id}`, {
+          method: 'DELETE',
+          headers: {
+            'Content-Type': 'application/json'
+          }
+        })
+          .then(res => res.json())
+          .then(data => {
+            console.log(data);
+            swal({
+              title: "Negociación eliminada con éxito",
+              icon: "success"
+            }).then(() => {
+              navegar(0);
+            });
+          })
+          .catch(error => {
+            console.error('Error:', error);
+            alert('Error al eliminar la negociación');
+          });
       }
-    })
-      .then(res => res.json())
-      .then(data => {
-        // console.log(data);
-        alert(data.message); // Mostrar el mensaje específico del objeto
-        navegar(0);
-      })
-      .catch(error => {
-        console.error('Error:', error);
-        alert('Error al eliminar la negociación');
-      });
-  }
+    });
+  };
 
   if (!negociacion) {
     return <div>No se ha proporcionado una negociación válida</div>;
@@ -102,14 +117,20 @@ const NegociacionIndividual = ({ negociacion }) => {
 
     const planDePago = [];
 
+    let fechaPago = new Date(fechaGracia); // Inicializar fechaPago con la fecha de gracia
+
     for (let i = 0; i < numCuotas; i++) {
-      const fechaPago = new Date(fechaGracia);
-      fechaPago.setDate(fechaPago.getDate() + 25 * i);
+      // Para la primera cuota, restar 5 días a la fecha de gracia
+      if (i === 0) {
+        fechaPago.setDate(fechaPago.getDate() - 5);
+      } else {
+        fechaPago.setDate(fechaPago.getDate() + 25); // Sumar 25 días a partir de la primera cuota
+      }
 
       planDePago.push({
         numCuota: i + 1, // Número de cuota
         fecha: format(fechaPago, 'dd/MM/yyyy'),
-        valor: valorCuota.toFixed(2),
+        valor: valorCuota.toLocaleString('es-CO', { minimumFractionDigits: 0 })
       });
     }
 
@@ -117,6 +138,8 @@ const NegociacionIndividual = ({ negociacion }) => {
   };
 
   const planPagoData = calcularPlanPago();
+
+  console.log(negociacion)
 
   return (
     <tr>
@@ -129,7 +152,7 @@ const NegociacionIndividual = ({ negociacion }) => {
       </td>
       <td>{negociacion.numCuotas}</td>
       <td>{fechaFormateada}</td>
-      <td>${negociacion.total}</td>
+      <td>$ {parseFloat(negociacion.total).toLocaleString('es-CO')}</td>
       <td style={{ textAlign: 'center' }}>
         <Link onClick={toggleDetalles} >
           <i className="fa fa-circle-info" title="Detalle" style={{ marginRight: 10, color: '#212529', fontSize: 22 }} />
@@ -147,119 +170,122 @@ const NegociacionIndividual = ({ negociacion }) => {
           <i className="fa fa-money" title="Plan de pago" style={{ color: '#212529', fontSize: 22 }} />
         </Link>
       </td>
-      {/* Modal */}
-      <Modal isOpen={mostrarDetalles} onRequestClose={toggleDetalles} style={customStyles} >
-        <Link onClick={closeModal}>
-          <FaTimes size={35} style={{ color: 'black', float: 'right' }} />
-        </Link>
-        <br />
-        <h2 style={{ textAlign: 'center' }}>Detalle Negociación</h2>
-        <br />
-        <table className="table table-hover mb-5 table-bordered" style={{ maxWidth: 800, border: "2px solid black" }}>
-          <tbody>
-            <tr>
-              <th scope="row" style={{ backgroundColor: "lightgray" }}>Cliente</th>
-              <td>{negociacion.cliente}</td>
-            </tr>
-            <tr>
-              <th scope="row" style={{ backgroundColor: "lightgray" }}>Factura</th>
-              <td>{negociacion.numFactura}</td>
-            </tr>
-            <tr>
-              <th scope="row" style={{ backgroundColor: "lightgray" }}>Productos</th>
-              <td dangerouslySetInnerHTML={{ __html: Array.isArray(negociacion.tipoMaquina) ? negociacion.tipoMaquina.join('<br />') : '' }}></td>
-            </tr>
-            <tr>
-              <th scope="row" style={{ backgroundColor: "lightgray" }}>Cantidad</th>
-              <td dangerouslySetInnerHTML={{ __html: Array.isArray(negociacion.cantidad) ? negociacion.cantidad.join('<br />') : '' }}></td>
-            </tr>
-            <tr>
-              <th scope="row" style={{ backgroundColor: "lightgray" }}>Precio Base</th>
-              <td dangerouslySetInnerHTML={{ __html: Array.isArray(negociacion.precioBase) ? negociacion.precioBase.join('<br />') : '' }}></td>
-            </tr>
-            <tr>
-              <th scope="row" style={{ backgroundColor: "lightgray" }}>Precio Venta</th>
-              <td dangerouslySetInnerHTML={{ __html: Array.isArray(negociacion.precioVenta) ? negociacion.precioVenta.join('<br />') : '' }}></td>
-            </tr>
-            <tr>
-              <th scope="row" style={{ backgroundColor: "lightgray" }}>Cuotas</th>
-              <td>{negociacion.numCuotas}</td>
-            </tr>
-            <tr>
-              <th scope="row" style={{ backgroundColor: "lightgray" }}>Tasa</th>
-              <td>{negociacion.tasa}</td>
-            </tr>
-            <tr>
-              <th scope="row" style={{ backgroundColor: "lightgray" }}>Anticipo</th>
-              <td>{negociacion.anticipo}</td>
-            </tr>
-            <tr>
-              <th scope="row" style={{ backgroundColor: "lightgray" }}>Intereses</th>
-              <td>{negociacion.interes}</td>
-            </tr>
-            <tr>
-              <th scope="row" style={{ backgroundColor: "lightgray" }}>Fecha Fin Gracia</th>
-              <td>{negociacion.fechaGracia}</td>
-            </tr>
-            <tr>
-              <th scope="row" style={{ backgroundColor: "lightgray" }}>Total</th>
-              <td>{negociacion.total}</td>
-            </tr>
-          </tbody>
-        </table>
-      </Modal>
-      {/* Modal 2 */}
+
+      {/* Modal Detalle Productos */}
       <Modal isOpen={showModal} onRequestClose={toggleModal} style={customStyles2}>
         <Link onClick={closeModal}>
           <FaTimes size={35} style={{ color: 'black', float: 'right' }} />
         </Link>
         <br />
-        <h2 style={{ textAlign: 'center' }}>Detalle Productos</h2>
+        <h2 style={{ textAlign: 'center', color: '#032770' }}>DETALLE PRODUCTOS</h2>
         <br />
         <div style={{ display: 'flex', justifyContent: 'center' }}>
-          <table className="table table-hover mb-5 table-bordered" style={{ maxWidth: 800 }}>
-            <thead className="table-secondary">
+          <table className="table table-hover mb-5 table-bordered" style={{ maxWidth: 800, border: "2px solid blue" }}>
+            <thead className="table-secondary" style={{ border: "2px solid blue" }}>
               <tr>
-                <th scope="col">Producto</th>
-                <th scope="col">Cantidad</th>
-                <th scope="col">Precio Base</th>
-                <th scope="col">Precio Venta</th>
+                <th scope="col" style={{ backgroundColor: "#032770", color: 'white' }}>Producto</th>
+                <th scope="col" style={{ backgroundColor: "#032770", color: 'white' }}>Cantidad</th>
+                <th scope="col" style={{ backgroundColor: "#032770", color: 'white' }}>Precio Base</th>
+                <th scope="col" style={{ backgroundColor: "#032770", color: 'white' }}>Precio Venta</th>
               </tr>
             </thead>
             <tbody>
               <tr>
-                <td style={{ textAlign: 'center' }} dangerouslySetInnerHTML={{ __html: negociacion.tipoMaquina.join('<br />') }}></td>
-                <td style={{ textAlign: 'center' }} dangerouslySetInnerHTML={{ __html: negociacion.cantidad.join('<br />') }}></td>
-                <td style={{ textAlign: 'center' }} dangerouslySetInnerHTML={{ __html: negociacion.precioBase.join('<br />') }}></td>
-                <td style={{ textAlign: 'center' }} dangerouslySetInnerHTML={{ __html: negociacion.precioVenta.join('<br />') }}></td>
+                <td style={{ color: '#032770' }} dangerouslySetInnerHTML={{ __html: negociacion.tipoMaquina.join('<br />') }}></td>
+                <td style={{ color: '#032770' }} dangerouslySetInnerHTML={{ __html: negociacion.cantidad.join('<br />') }}></td>
+                <td style={{ color: '#032770' }} dangerouslySetInnerHTML={{ __html: Array.isArray(negociacion.precioBase) ? negociacion.precioBase.map((precio) => `$ ${parseFloat(precio).toLocaleString('es-CO')}`).join('<br />') : '' }}></td>
+                <td style={{ color: '#032770' }} dangerouslySetInnerHTML={{ __html: Array.isArray(negociacion.precioVenta) ? negociacion.precioVenta.map((precio) => `$ ${parseFloat(precio).toLocaleString('es-CO')}`).join('<br />') : '' }}></td>
               </tr>
             </tbody>
           </table>
         </div>
       </Modal>
+
+      {/* Modal Detalle Negociacion */}
+      <Modal isOpen={mostrarDetalles} onRequestClose={toggleDetalles} style={customStyles} >
+        <Link onClick={closeModal}>
+          <FaTimes size={35} style={{ color: 'black', float: 'right' }} />
+        </Link>
+        <br />
+        <h2 style={{ textAlign: 'center', color: '#032770' }}>DETALLE NEGOCIACIÓN</h2>
+        <br />
+        <table className="table table-hover mb-5 table-bordered" style={{ maxWidth: 800, border: "2px solid blue" }}>
+          <tbody style={{ border: "2px solid blue" }}>
+            <tr>
+              <th scope="row" style={{ backgroundColor: "#032770", color: 'white' }}>Cliente</th>
+              <td style={{ color: '#032770' }}>{negociacion.cliente}</td>
+            </tr>
+            <tr>
+              <th scope="row" style={{ backgroundColor: "#032770", color: 'white' }}>Factura</th>
+              <td style={{ color: '#032770' }}>{negociacion.numFactura}</td>
+            </tr>
+            <tr>
+              <th scope="row" style={{ backgroundColor: "#032770", color: 'white' }}>Productos</th>
+              <td style={{ color: '#032770' }} dangerouslySetInnerHTML={{ __html: Array.isArray(negociacion.tipoMaquina) ? negociacion.tipoMaquina.join('<br />') : '' }}></td>
+            </tr>
+            <tr>
+              <th scope="row" style={{ backgroundColor: "#032770", color: 'white' }}>Cantidad</th>
+              <td style={{ color: '#032770' }} dangerouslySetInnerHTML={{ __html: negociacion.cantidad.join('<br />') }}></td>
+            </tr>
+            <tr>
+              <th scope="row" style={{ backgroundColor: "#032770", color: 'white' }}>Precio Base</th>
+              <td style={{ color: '#032770' }} dangerouslySetInnerHTML={{ __html: Array.isArray(negociacion.precioBase) ? negociacion.precioBase.map((precio) => `$ ${parseFloat(precio).toLocaleString('es-CO')}`).join('<br />') : '' }}></td>
+            </tr>
+            <tr>
+              <th scope="row" style={{ backgroundColor: "#032770", color: 'white' }}>Precio Venta</th>
+              <td style={{ color: '#032770' }} dangerouslySetInnerHTML={{ __html: Array.isArray(negociacion.precioVenta) ? negociacion.precioVenta.map((precio) => `$ ${parseFloat(precio).toLocaleString('es-CO')}`).join('<br />') : '' }}></td>
+            </tr>
+            <tr>
+              <th scope="row" style={{ backgroundColor: "#032770", color: 'white' }}>Cuotas</th>
+              <td style={{ color: '#032770' }}>{negociacion.numCuotas}</td>
+            </tr>
+            <tr>
+              <th scope="row" style={{ backgroundColor: "#032770", color: 'white' }}>Tasa</th>
+              <td style={{ color: '#032770' }}>{negociacion.tasa}</td>
+            </tr>
+            <tr>
+              <th scope="row" style={{ backgroundColor: "#032770", color: 'white' }}>Anticipo</th>
+              <td style={{ color: '#032770' }}>{negociacion.anticipo}</td>
+            </tr>
+            <tr>
+              <th scope="row" style={{ backgroundColor: "#032770", color: 'white' }}>Intereses</th>
+              <td style={{ color: '#032770' }}>{negociacion.interes}</td>
+            </tr>
+            <tr>
+              <th scope="row" style={{ backgroundColor: "#032770", color: 'white' }}>Fecha Fin Gracia</th>
+              <td style={{ color: '#032770' }}>{fechaFormateada}</td>
+            </tr>
+            <tr>
+              <th scope="row" style={{ backgroundColor: "#032770", color: 'white' }}>Total</th>
+              <td style={{ color: '#032770' }}>$ {negociacion.total.toLocaleString('es-CO', { minimumFractionDigits: 0 })}</td>
+            </tr>
+          </tbody>
+        </table>
+      </Modal>
+
       {/* Modal Planes de Pago */}
       <Modal isOpen={showPlanPagoModal} onRequestClose={togglePlanPagoModal} style={customStyles2}>
         <Link onClick={closeModalPlanPago}>
           <FaTimes size={35} style={{ color: 'black', float: 'right' }} />
         </Link>
         <br />
-        <h2 style={{ textAlign: 'center' }}>Plan de Pago</h2>
+        <h2 style={{ textAlign: 'center', color: "#032770" }}>PLAN DE PAGO</h2>
         <br />
         <div style={{ display: 'flex', justifyContent: 'center' }}>
-          <table className="table table-hover mb-5 table-bordered" style={{ maxWidth: 800 }}>
-            <thead className="table-secondary">
+          <table className="table table-hover mb-5 table-bordered" style={{ maxWidth: 800, border: "2px solid blue" }}>
+            <thead className="table-secondary" style={{ border: "2px solid blue" }}>
               <tr>
-                <th scope="col">Cuota</th>
-                <th scope="col">Fecha</th>
-                <th scope="col">Valor</th>
+                <th scope="col" style={{ backgroundColor: "#032770", color: 'white' }}>Cuota</th>
+                <th scope="col" style={{ backgroundColor: "#032770", color: 'white' }}>Fecha</th>
+                <th scope="col" style={{ backgroundColor: "#032770", color: 'white' }}>Valor</th>
               </tr>
             </thead>
             <tbody>
               {planPagoData.map((item, index) => (
                 <tr key={index}>
-                  <td>{item.numCuota}</td>
-                  <td>{item.fecha}</td>
-                  <td>${item.valor}</td>
+                  <td style={{ color: '#032770' }}>{item.numCuota}</td>
+                  <td style={{ color: '#032770' }}>{item.fecha}</td>
+                  <td style={{ color: '#032770' }}>$ {item.valor}</td>
                 </tr>
               ))}
             </tbody>
